ServiceSLO value and target are now floats

+++ dashdotdb/models/dashdotdb.py
@@ -171,8 +171,8 @@ class ServiceSLO(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(64), unique=False)
-    value = db.Column(db.Integer, unique=False)
-    target = db.Column(db.Integer, unique=False)
+    value = db.Column(db.Float, unique=False)
+    target = db.Column(db.Float, unique=False)
     slitype_id = db.Column(db.Integer, db.ForeignKey('slitype.id'))
     token_id = db.Column(db.Integer, db.ForeignKey('token.id'))
     service_id = db.Column(db.Integer, db.ForeignKey('service.id'))
